chore: move src/utils/helpers to src/utils/preview
